Add ParseIndex #398
Ineteger variable in the POSaM is set to int64 as default, for the
reason of scientific computing. That's why all the type assertion is
specified by int64. But for the index, indicating the elements in array,
it should be int, so at the end of the function, an int converting is
necessary