Swap printhead offset and add slide width and height #408
Not all the height is available for the current motion flatform.

the sum of printhead offset and slide height should be less than 50.0,
whatever the direction of the offset, since the offset is moving to the
origin point. e.g., offset -20 to the origin, then the max height of
moving upward(same direction) is 30, and offset 20 is the same.